chore: experiment with client components
