+++ client/app/operations/page.tsx
@@ -1,10 +1,14 @@
-"use client";
-
 import { GridRowsProp, GridColDef } from "@mui/x-data-grid";
 import DataGrid from "../components/DataGrid/DataGrid";
 import Link from "next/link";
 import { Button } from "@mui/material";
-import { useGetOperationsQuery } from "@/redux/index";
+import { Suspense } from "react";
+
+export async function getOperations() {
+  return (
+    await fetch("http://localhost:8000/api/registration/operations")
+  ).json();
+}
 
 export const columns: GridColDef[] = [
   { field: "operation_id", headerName: "Operation ID", width: 150 },
@@ -18,34 +22,25 @@ export const columns: GridColDef[] = [
     headerName: "Action",
     sortable: false,
     width: 200,
-    renderCell: (params) => {
-      return (
-        <Link href={`operations/${params.row.id}`}>
-          <Button variant="contained">
-            {params.row.status === "Not Registered"
-              ? "Start Registration"
-              : "View Details"}
-          </Button>
-        </Link>
-      );
-    },
+    // brianna figure out
+    // renderCell: (params) => {
+    //   return (
+    //     <Link href={`operations/${params.row.id}`}>
+    //       <Button variant="contained">
+    //         {params.row.status === "Not Registered"
+    //           ? "Start Registration"
+    //           : "View Details"}
+    //       </Button>
+    //     </Link>
+    //   );
+    // },
   },
 ];
 
-export default function Page() {
-  const { data, isLoading, error } = useGetOperationsQuery(null);
+export default async function Page() {
+  const operations = await getOperations();
 
-  // Check if data is currently loading
-  if (isLoading) {
-    return <div>🚀 Loading data... </div>;
-  }
-
-  // Check if an error occurred
-  if (error) {
-    return <div>Something went wrong, Please retry after some time</div>;
-  }
-
-  if (!data) {
+  if (!operations) {
     return (
       <div>
         No operations data in database (did you forget to run `make
@@ -54,7 +49,7 @@ export default function Page() {
     );
   }
 
-  const rows: GridRowsProp = data.map(
+  const rows: GridRowsProp = operations.map(
     ({
       id,
       registration_year,
@@ -78,10 +73,12 @@ export default function Page() {
   return (
     <>
       <h1>Operations List</h1>
-      <Link href="/operations/create">
-        <Button variant="contained">Add Operation</Button>
-      </Link>
-      <DataGrid columns={columns} rows={rows} />
+      <Suspense fallback={<p>🚀 Loading data...</p>}>
+        <Link href="/operations/create">
+          <Button variant="contained">Add Operation</Button>
+        </Link>
+        <DataGrid columns={columns} rows={rows} />
+      </Suspense>
     </>
   );
 }
